+++ src/main/java/insane96mcp/runeenchanting/setup/RERunes.java
@@ -144,6 +144,7 @@ public class RERunes {
     public static final DeferredHolder<Rune, AtmosphericRune> ATMOSPHERIC = register("atmospheric", new AtmosphericRune());
     public static final DeferredHolder<Rune, LapisFueledRune> LAPIS_FUELED = register("lapis_fueled", new LapisFueledRune());
     public static final DeferredHolder<Rune, BloodFueledRune> BLOOD_FUELED = register("blood_fueled", new BloodFueledRune());
+    public static final DeferredHolder<Rune, SoulboundRune> SOULBOUND = register("soulbound", new SoulboundRune());
 
     // Curse
     public static final DeferredHolder<Rune, CurseOfBindingRune> CURSE_OF_BINDING = register("curse_of_binding", new CurseOfBindingRune());